+++ R/geom-quiver.r
@@ -55,7 +55,11 @@ geom_quiver <- function(mapping = NULL, data = NULL,
 #' @export
 GeomQuiver <- ggproto(
   "GeomQuiver", ggplot2::GeomSegment,
-  draw_panel = function(data, panel_params, coord, arrow = NULL, lineend = "butt", na.rm = FALSE) {
+  draw_panel = function(
+    data, panel_params, coord, 
+    arrow = grid::arrow(), 
+    lineend = "butt", na.rm = FALSE
+  ) {
     # Apply coordinate transformations to get proper arrow lengths
     if(inherits(coord, "CoordMap")) {
       # Workaround for CoordMap transform method not transforming xend and yend
@@ -67,12 +71,13 @@ GeomQuiver <- ggproto(
     } else {
       t_data <- coord$transform(data, panel_params)
     }
-    data$arrowsize <- with(t_data, sqrt((x - xend) ^ 2 + (y - yend) ^ 2) * 0.5)
+    
+    arrow$length <- arrow$length*10*with(t_data, sqrt((x - xend) ^ 2 + (y - yend) ^ 2) * 0.5)
+
     # Re-use segments to produce arrows
     ggplot2::GeomSegment$draw_panel(
       data, panel_params, coord,
-      arrow = grid::arrow(length = unit(data$arrowsize, "npc")),
-      lineend = lineend
+      arrow = arrow, lineend = lineend
     )
   }
 )