+++ tests/testthat/test-geom_quiver.R
@@ -30,3 +30,15 @@ test_that("Simple trig quiver plot", {
     ggplot(aes(x = x, y = y, u = u, v = v)) +
     geom_quiver()
 })
+
+test_that("Custom arrows with grid::arrow", {
+  library(ggplot2)
+  plotdata <- expand.grid(x = seq(0, pi, pi / 12), y = seq(0, pi, pi / 12))
+  plotdata$u <- cos(plotdata$x)
+  plotdata$v <- sin(plotdata$y)
+
+  p <- plotdata |>
+    ggplot(aes(x = x, y = y, u = u, v = v)) +
+    geom_quiver(arrow = grid::arrow(type = "closed"))
+  vdiffr::expect_doppelganger("quiver plot with closed arrow", p)
+})